test: 게시글 상세 조회하는 커스텀 훅에 대해서 테스트 한다
